requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
googlemaps==4.10.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.4
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.1
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
